添加了项目修改，小组修改，任务修改等接口

+++ src/main/java/org/xm06/pms/utils/FileUtil.java
@@ -152,11 +152,13 @@ public class FileUtil {
             return;
         }
         File[] files = file.listFiles();
-        for (File f: files){
-            if (f.isDirectory()){
-                deleteFile(f);
-            }else {
-                f.delete();
+        if(files != null){
+            for (File f: files){
+                if (f.isDirectory()){
+                    deleteFile(f);
+                }else {
+                    f.delete();
+                }
             }
         }
         file.delete();
